+++ PKHeX.Core/Legality/Encounters/Generator/Search/EncounterEnumerator8b.cs
@@ -165,7 +165,7 @@ public record struct EncounterEnumerator8b(PKM Entity, EvoCriteria[] Chain, Game
         var remap = LocationsHOME.GetRemapState(EntityContext.Gen8b, Entity.Context);
         hasOriginalLocation = true;
         if (remap.HasFlag(LocationRemapState.Remapped))
-            hasOriginalLocation = location != LocationsHOME.GetMetSWSH((ushort)location, (int)Version);
+            hasOriginalLocation = location != LocationsHOME.GetMetSWSH((ushort)location, Version);
     }
 
     private bool TryGetNext<TArea, TSlot>(TArea[] areas)
