+++ PKHeX.Core/Legality/Encounters/Information/EncounterSuggestionData.cs
@@ -39,5 +39,5 @@ public sealed class EncounterSuggestionData : ISpeciesForm, IRelearn, ILevelRang
 
     public int GetSuggestedMetLevel(PKM pk) => EncounterSuggestion.GetSuggestedMetLevel(pk, LevelMin);
     public GroundTileType GetSuggestedGroundTile() => Encounter is IGroundTypeTile t ? t.GroundTile.GetIndex() : 0;
-    public bool HasGroundTile(int format) => Encounter is IGroundTypeTile t && t.HasGroundTile(format);
+    public bool HasGroundTile(byte format) => Encounter is IGroundTypeTile t && t.HasGroundTile(format);
 }
