+++ PKHeX.Core/Legality/LearnSource/Group/LearnGroup3.cs
@@ -8,7 +8,7 @@ namespace PKHeX.Core;
 public sealed class LearnGroup3 : ILearnGroup
 {
     public static readonly LearnGroup3 Instance = new();
-    private const int Generation = 3;
+    private const byte Generation = 3;
     public ushort MaxMoveID => Legal.MaxMoveID_3;
 
     public ILearnGroup? GetPrevious(PKM pk, EvolutionHistory history, IEncounterTemplate enc, LearnOption option) => null; // Gen3 is the end of the line!
