+++ PKHeX.Core/Legality/LearnSource/Group/LearnGroup9.cs
@@ -8,7 +8,7 @@ namespace PKHeX.Core;
 public sealed class LearnGroup9 : ILearnGroup
 {
     public static readonly LearnGroup9 Instance = new();
-    private const int Generation = 9;
+    private const byte Generation = 9;
     private const EntityContext Context = EntityContext.Gen9;
     public ushort MaxMoveID => Legal.MaxMoveID_9;
