+++ PKHeX.Core/Moves/MoveInfo.cs
@@ -230,7 +230,7 @@ public static class MoveInfo
         return types[move];
     }
 
-    public static bool IsAnyFromGeneration(int generation, ReadOnlySpan<MoveResult> moves)
+    public static bool IsAnyFromGeneration(byte generation, ReadOnlySpan<MoveResult> moves)
     {
         foreach (var move in moves)
         {
