+++ PKHeX.Core/MysteryGifts/MysteryGift.cs
@@ -80,7 +80,7 @@ public abstract class MysteryGift : IEncounterable, IMoveset, IRelearn, ITrainer
 
     public string Extension => GetType().Name.ToLowerInvariant();
     public string FileName => $"{CardHeader}.{Extension}";
-    public abstract int Generation { get; }
+    public abstract byte Generation { get; }
     public abstract EntityContext Context { get; }
     public abstract bool FatefulEncounter { get; }
 
@@ -159,7 +159,7 @@ public abstract class MysteryGift : IEncounterable, IMoveset, IRelearn, ITrainer
     public virtual bool IsEgg { get => false; set { } }
     public virtual int HeldItem { get => -1; set { } }
     public virtual int AbilityType { get => -1; set { } }
-    public abstract int Gender { get; set; }
+    public abstract byte Gender { get; set; }
     public abstract byte Form { get; set; }
     public abstract uint ID32 { get; set; }
     public abstract ushort TID16 { get; set; }
