+++ PKHeX.Core/PKM/Interfaces/IAwakened.cs
@@ -233,7 +233,7 @@ public static class AwakeningUtil
     {
         // GO Park transfers start with 2 AVs for all stats.
         // Every other encounter is either all 0, or can legally start at 0 (trades).
-        if (pk.Version == (int)GameVersion.GO)
+        if (pk.Version == GameVersion.GO)
             result.Fill(GP1.InitialAV);
 
         // Leveling up in-game adds 1 AV to a "random" index.
